Create radix-2_multiplier.sv
This a design a radix-2 multiplier with shift and add algorithm. This multiplier works on every bit and hence takes 32 cycles of operation.

Multiplier and Multiplicand is stored in separate registers. a A register double the width the operands is used to store the result from the adder sum and carry. Addition is done on the positive edge of the clock and shifting is done on the negative edge of the clock thereby completing the operation for one bit in a two sub-cycles.